Consolidate model classes and get generate working
